Changed creation of edges so that the arrows are displayed right from the start. I'm not sure if this is a zest bug or not, but it only works if the CONNECTIONS_DIRECTED style is part of the style bitmask used when creating the GraphConnection.
Improved filter mechanism: Consider an edge that has no category assigned, but
links two nodes wich have. If one of these nodes were filtered, this lead to
Exceptions. ==> An edge is now only part of the graph if both of its nodes are.
If one of its nodes is suppressed because of the category filter, the edge will
be suppressed, too.

git-svn-id: http://emfmodelvisualizer.googlecode.com/svn@58 d18ea948-b79e-11dd-b040-2d45f096adfe