Add ISelectionProvider / PropertySource support
The ModelViewer now implements ISelectionProvider and forwards selections from contained graphs.
An adapter factory which adapts graphmm EObjects to IPropertySource has been added.

The combination of these two changes allows us to display model data in Eclipse's Properties view. It usually has a lot more (horizontal) sopace as the Table which is currently used to display a model object's properties.

We might decide to drop the properties table or make it optional. But for now both views are active.

git-svn-id: http://emfmodelvisualizer.googlecode.com/svn@122 d18ea948-b79e-11dd-b040-2d45f096adfe